+++ src/org/cicirello/search/operators/permutations/WindowLimitedReversalMutation.java
@@ -103,7 +103,11 @@ public final class WindowLimitedReversalMutation extends ReversalMutation {
 	
 	@Override
 	final void generateIndexes(int n, int[] indexes) {
-		RandomIndexer.nextWindowedIntPair(n, limit, indexes);
+		if (limit >= n) {
+			super.generateIndexes(n, indexes);
+		} else {
+			RandomIndexer.nextWindowedIntPair(n, limit, indexes);
+		}
 	}
 	
 }