+++ src/org/cicirello/search/operators/permutations/WindowLimitedScrambleMutation.java
@@ -78,7 +78,11 @@ public final class WindowLimitedScrambleMutation extends ScrambleMutation {
 	
 	@Override
 	final void generateIndexes(int n, int[] indexes) {
-		RandomIndexer.nextWindowedIntPair(n, limit, indexes);
+		if (limit >= n) {
+			super.generateIndexes(n, indexes);
+		} else {
+			RandomIndexer.nextWindowedIntPair(n, limit, indexes);
+		}
 	}
 
 }