+++ src/org/cicirello/search/operators/permutations/WindowLimitedSwapMutation.java
@@ -82,6 +82,10 @@ public final class WindowLimitedSwapMutation extends SwapMutation {
 	
 	@Override
 	final void generateIndexes(int n, int[] indexes) {
-		RandomIndexer.nextWindowedIntPair(n, limit, indexes);
+		if (limit >= n) {
+			super.generateIndexes(n, indexes);
+		} else {
+			RandomIndexer.nextWindowedIntPair(n, limit, indexes);
+		}
 	}
 }